requirements.txt
oryks-youtube
langchain
crewai